requirements.txt
flask-restplus
flask-restx
scikit-learn==1.5.0
joblib
numpy


